% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHyperlinks.R
\docType{methods}
\name{addHyperLinks}
\alias{addHyperLinks}
\alias{addHyperLinks,jobjRef-method}
\title{Add Hyper Links}
\usage{
addHyperLinks(sheet, col.name, urls)

\S4method{addHyperLinks}{jobjRef}(sheet, col.name, urls)
}
\arguments{
\item{sheet}{An \pkg{xlsx} or \pkg{xksxSimple}, worksheet.}

\item{col.name}{A character scalar, the name of the column where the hyperlinks are to be added. The column will be created if it doesn't exist.}

\item{urls}{A character vector, the list of the hyperlink targets.}
}
\value{
The worksheet invisibly.
}
\description{
\code{addHyperlinks} adds hyperlinks to an \pkg{xlsx} or \pkg{xksxSimple}, Excel worksheet.
}
\examples{
\dontrun{
toSheet(mtcars) -> sheet
addHyperLinks(sheet, "Google", "http://google.com")
}

}
