% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkbook.R
\name{createWorkbook}
\alias{createWorkbook}
\title{Create an Empty Workbook object}
\usage{
createWorkbook(type = "xlsx")
}
\arguments{
\item{...}{Arguments to pass through to \code{xlsx::createWorkbook}}
}
\value{
The worksheet object.
}
\description{
Create an Empty Workbook object
}
\examples{
\dontrun{
createWorkbook() -> book
}

}
\seealso{
\code{\link[xlsx]{createWorkbook}}
}
