% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnValue.R
\docType{methods}
\name{getColumnValue}
\alias{getColumnValue}
\alias{getColumnValue,jobjRef-method}
\title{Get Column Value}
\usage{
getColumnValue(sheet, col.name)

\S4method{getColumnValue}{jobjRef}(sheet, col.name)
}
\arguments{
\item{sheet}{An xlsx or xlsxsimple worksheet.}

\item{col.name}{A character scalar, the name of the column to return.}
}
\value{
A vector.
}
\description{
Gets the values from column in an xlsx or xlsxsimple, Excel worksheet.
}
\examples{
\dontrun{
sheet <- toSheet(mtcars)
getColumnValue(sheet, "mpg")
}

}
