% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveWorkbook.R
\name{saveWorkbook}
\alias{saveWorkbook}
\title{Save Work Book}
\usage{
saveWorkbook(workbook = getWorkbook(), file.path = thisFile(), title = NA,
  subject = NA, creator = Sys.getenv("USERNAME"), start = FALSE)
}
\arguments{
\item{workbook}{An optional xlsxSimple workbook.}

\item{file.path}{An optional character scalar, the path and filename for the file.}

\item{title}{An optional character scalar, the title of the workbook.}

\item{subject}{An optional character scalar, the subject of the workbook.}

\item{creator}{An optional character scalar, the creator of the workbook. Defaults to the logged-on user name.}

\item{start}{An optional logical scalar, should the workbook be started in Excel.}
}
\value{
A character scalar, the path of the file, invisibly.
}
\description{
\code{saveWorkbook} saves an Excel workbook on the file system, setting the metadata. Optionally, the file is opened in Excel.
}
\note{
If the \code{file.path} does not end with ``.xlsx'' then the fiile extension will be set to ``.xlsx''. If \code{file,path} is omitted then the script file path will be used, but that extension substiuted. But an error it thrown if \code{file.path} is omitted in interactive mode.
}
\examples{

book <- createWorkbook()
saveWorkbook(book, "foo.xlsx", "Foo", "Examples", start=TRUE)

}
