% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toExcel.R
\name{toExcel}
\alias{toExcel}
\title{toExcel}
\usage{
toExcel(x, filePath = thisFile(), title = NA, subject = NA,
  creator = Sys.getenv("USERNAME"), start = TRUE, ...)
}
\arguments{
\item{x}{A data frame, plot or path to a saved plot.}

\item{filePath}{An optional character scalar, the path and filename for the file.}

\item{title}{An optional character scalar, the title of the workbook.}

\item{subject}{An optional character scalar, the subject of the workbook.}

\item{creator}{An optional character scalar, the creator of the workbook. Defaults to the logged-on user name.}

\item{start}{An optional logical scalar, should the workbook be started in Excel.}
}
\value{
A character scalar, the path of the file, invisibly.
}
\description{
Save something to a single sheet, Excel workbook. This is for when calling \code{toSheet()} and \code{saveWorkbook()} is too much work. Which happens when one is developing something.
}
\examples{

toExcel(mtcars)

}
\seealso{
See \code{\link{toSheet}} and \code{\link{saveWorkbook}} for details. This function calls and then the other.
}
